+++ System/storage/Storage_Dev_Port.c
@@ -52,9 +52,9 @@ static uint8_t write_tmp[Storage_TabSize * 2] __attribute__((aligned(4))) __attr
 static bool Storage_Dev_Set(StorageDevObj_TypeDef *ext_dev);
 static bool Storage_Dev_Init(StorageDevObj_TypeDef *ext_dev, uint32_t *p_code);
 
-static bool Storage_Dev_Write_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint8_t *p_data, uint16_t len);
-static bool Storage_Dev_Read_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint8_t *p_data, uint16_t len);
-static bool Storage_Dev_Erase_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint16_t len);
+static bool Storage_Dev_Write_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint8_t *p_data, uint32_t len);
+static bool Storage_Dev_Read_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint8_t *p_data, uint32_t len);
+static bool Storage_Dev_Erase_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint32_t len);
 
 static bool Storage_Dev_Param_Read(StorageDevObj_TypeDef *p_dev, uint32_t base_addr, uint8_t *p_data, uint32_t len);
 static bool Storage_Dev_Param_Write(StorageDevObj_TypeDef *p_dev, uint32_t base_addr, uint8_t *p_data, uint32_t len);
@@ -154,7 +154,7 @@ static bool Storage_Dev_Init(StorageDevObj_TypeDef *ext_dev, uint32_t *p_code)
     return (init_state == 0) ? true : false;
 }
 
-static bool Storage_Dev_Write_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint8_t *p_data, uint16_t len)
+static bool Storage_Dev_Write_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint8_t *p_data, uint32_t len)
 {
     uint32_t write_cnt = 0;
     uint32_t addr_tmp = 0;
@@ -185,7 +185,7 @@ static bool Storage_Dev_Write_Section(StorageDevObj_TypeDef *p_dev, uint32_t add
     return true;
 }
 
-static bool Storage_Dev_Read_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint8_t *p_data, uint16_t len)
+static bool Storage_Dev_Read_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint8_t *p_data, uint32_t len)
 {
     uint32_t read_cnt = 0;
     uint32_t addr_tmp = 0;
@@ -217,7 +217,7 @@ static bool Storage_Dev_Read_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr
     return true;
 }
 
-static bool Storage_Dev_Erase_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint16_t len)
+static bool Storage_Dev_Erase_Section(StorageDevObj_TypeDef *p_dev, uint32_t addr, uint32_t len)
 {
     uint32_t erase_cnt = 0;
     uint32_t addr_tmp = 0;
